Improve performance
Issue: When fast clicking between pages (i.e. Home and Privacy) I noticed that the CSS gets applied after the page is rendered by the browser. That’s because the npm package style-loader plugs the CSS into the page after it was loaded causing the browser to re-render the page!
On top of that, the Javascript bundle — which contains the CSS — is very large. It contains the entire Bootstrap CSS as well as some Bootstrap Javascript functions and all of JQuery!

Fix:
Bundling CSS into a separate CSS file that can be statically referenced from an HTML link tag
Getting rid of all the CSS that we’re not using
Splitting up the Javascript code into separate bundles. Many pages have Javascript code that is unique to them and not every page needs JQuery

In the standard MVC template all of Bootstrap and JQuery are referenced and downloaded by the browser.

Validation scripts (basically JQuery with some extras): for forms that use ASP.NET Core’s form validation have been added to certain Views that have validation in the Model/ViewModels.

ADDED 4 javaScript files
site.js: Javascript and CSS code that is needed on each page (currently contains Bootstrap CSS and custom CSS). Will result in a separate JS and CSS file as seen in the diagram
bootstrap_js.js: Bootstrap’s Javascript code
validation.js: JQuery, including the validation scripts for our forms
index.js: some dummy code that’s only applicable to ‘Home’.

I now have a flexible and extensible project that is using modern front-end technologies and has excellent performance.